Los items funcionan correctamente, nuevas mejoras en la ventana de simulacion y se separa la clase principal del archivo main.py
